Split tests into smaller files (#1360)
This PR splits the tests/tests.rs files into smaller files
along the submodules in tests.rs.

This makes handling tests easier.
