+++ src/internet_identity/tests/integration/active_anchor_stats/mod.rs
@@ -1,0 +1,7 @@
+//! Module for active anchor statistics
+
+/// Tests for the general (domain independent) active anchor statistics  
+mod all_domains;
+
+/// Tests for the active anchor statistics that are specific to the II domains.
+mod ii_domains;
