+++ src/internet_identity/tests/integration/anchor_management/mod.rs
@@ -1,0 +1,6 @@
+//! Tests for the anchor management functionality of the Internet Identity.
+
+mod device_management;
+mod last_usage_timestamp;
+mod registration;
+mod remote_device_registration;
